logs relevant data under logs folder #13
