+++ lib/index.js
@@ -2,10 +2,8 @@ const express = require('express')
 const bodyParser = require('body-parser')
 const cors = require('cors')
 const morgan = require('morgan')
-const { clone } = require('./services/clone.js')
-const { build } = require('./services/compile')
-const { setPending, setSuccess, setFailure } = require('./services/status.js')
-const fs = require('fs');
+const { ci } = require('./services/ci.js')
+
 
 
 const PORT = 3000
@@ -27,40 +25,7 @@ app.post('/event', (req, res) => {
   res.sendStatus(201)
 })
 
-async function ci (req) {
-  console.log(req.body.repository.git_url)
-  console.log(req.body.pull_request.head.sha)
-  try{
-    ans = await clone(req.body.repository.clone_url, req.body.pull_request.head.ref, req.body.pull_request.head.sha, req.body.repository.full_name)
-    if(!ans){ 
-    }
-    let response = build('tmp/', req.body.pull_request.head.sha)
-    console.log(response)
-    if(!response.build_success){
-      setFailure(req.body.repository.full_name, req.body.pull_request.head.sha)
-      return
-    }
-    else{
-      setSuccess(req.body.repository.full_name, req.body.pull_request.head.sha)
-      writeJSON(req)
-      return
-    }    
-  }
-  catch (error){
-    console.log(error)
-  }
-  //console.log(buildResponse)
-}
-
 app.listen(PORT, () => {
   console.log(`Server listening on port ${PORT}`)
 })
 
-function writeJSON(req){
-  console.log("Writing JSON file")
-  fs.writeFile("requestData.json", JSON.stringify(req.body.pull_request), function(err) {
-    if (err) {
-        console.log(err);
-    }
-  });
-}
