+++ lib/services/clone.js
@@ -27,6 +27,9 @@ async function clone (URL, branch, commitID, repo) {
       shell.rm('-Rf', 'lib/services/tmp_Clone/{*,.*}')
       shell.rm('-R', 'lib/services/tmp_Clone')
 
+      //  Create log folder
+      shell.mkdir('-p', 'logs/' + commitID)
+
       //  Display errors
     }).catch(function (e) {
       status.setFailure(repo, commitID)
